pyproject.toml
[project]
name = "calorie_bot"
version = "0.1.0"
dependencies = [
    "python-telegram-bot",
    "matplotlib",
    "apscheduler",
    "redis",
    "ollama",
    "deep-translator",
    "sqlite-utils",
    "dotenv>=0.9.9",
]
requires-python = ">=3.12"
description = "A Telegram bot for tracking calorie intake and providing nutritional information. (Inspired by Salvatore Sanfilippo's work)"
readme = "README.md"
license = { text = "MIT" }
